backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "openai-agents>=0.1.0",
    "sqlmodel>=0.0.22",
    "asyncpg>=0.30.0",
    "alembic>=1.18.4",
    "fastembed>=0.4.0",
    "pgvector>=0.3.0",
    "python-dotenv>=1.0.0",
    "vaderSentiment>=3.3.2",
    "uvicorn>=0.32.0",
    "aiokafka>=0.13.0",
    "google-cloud-pubsub>=2.35.0",
    "twilio>=9.10.2",
    "redis>=7.2.1",
    "tenacity>=9.1.4",
    "email-validator>=2.3.0",
    "prometheus-client>=0.24.1",
    "google-api-python-client>=2.191.0",
    "google-auth-httplib2>=0.3.0",
    "google-auth-oauthlib>=1.3.0",
    "aiosqlite>=0.22.1",
]

[project.scripts]
main = "src.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
filterwarnings = [
    "ignore::pydantic.warnings.PydanticDeprecatedSince20"
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "httpx>=0.28.0",
]

